requirements.txt
version: '3.8'

services:
  discord-bot:
    build: .
    container_name: discord-bot
    restart: always
    environment:
      - DISCORD_BOT_TOKEN=your_bot_token_here
      - DISCORD_CHANNEL_ID=your_channel_id_here
